Move generator function to utils module
